refactor category page
